requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
hypothesis>=6.0.0
pytest>=8.0.0
